requirements.txt
openai
tiktoken
python-dotenv
deep_translator
whisper~=1.1.10
requests~=2.32.3
gTTS~=2.5.4
pydub~=0.25.1
colorama~=0.4.6
yt_dlp